fixup! s/64/32

+++ test/built-ins/DataView/prototype/getFloat32/minus-zero.js
@@ -37,6 +37,6 @@ sample.setUint8(1, 0);
 sample.setUint8(2, 0);
 sample.setUint8(3, 0);
 
-var result = sample.getFloat64(0);
+var result = sample.getFloat32(0);
 assert.sameValue(result, 0);
 assert.sameValue(1 / result, -Infinity);
